requirements.txt
ollama
rich
typer[all]
prompt-toolkit
pytest
pytest-asyncio
pyperclip
google-generativeai
anthropic
textual
